+++ Deeploy/Targets/PULPOpen/Templates/ReshapeTemplate.py
@@ -2,9 +2,6 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from typing import Dict, List, Tuple
-
-from Deeploy.DeeployTypes import NetworkContext, OperatorRepresentation, VariableBuffer
 from Deeploy.Targets.Generic.Templates.ReshapeTemplate import _ReshapeTemplate as _GenericReshapeTemplate
 
 
@@ -12,24 +9,6 @@ class _ReshapeTemplate(_GenericReshapeTemplate):
 
     def __init__(self, templateStr):
         super().__init__(templateStr)
-
-    def alignToContext(self, ctxt: NetworkContext,
-                       operatorRepresentation: OperatorRepresentation) -> Tuple[NetworkContext, Dict, List[str]]:
-
-        ctxt, operatorRepresentation, _ = super().alignToContext(ctxt, operatorRepresentation)
-
-        # Get buffers
-        bufferIn = ctxt.lookup(operatorRepresentation['data_in'])
-        assert isinstance(bufferIn, VariableBuffer)
-
-        bufferOut = ctxt.lookup(operatorRepresentation['data_out'])
-        assert isinstance(bufferOut, VariableBuffer)
-
-        # HACK: Tiling wasn't updated in the Fix aliasing PR so we have to still
-        #       set the _alias argument
-        bufferOut._alias = bufferIn.name
-
-        return ctxt, operatorRepresentation, []
 
 
 referenceTemplate = _ReshapeTemplate("""